wrapped library-generated graph with my own objects following red blob games tutorial
This added a bunch of complications regarding constructing edges and ordering corners which the library had been handling for me. Oh well!
Currently my solutions to this appear to add significant performance overhead to the process, so it would be worth reassessing whether duplicating the structure myself is necessary or whether I can work more within the library's structures.