increased world generation efficiency by a couple of orders of magnitude by trading off border geometry against big-O considerations.
Old method used voronoi regions, which broke the 1-1 mapping of edge vertices to results vertices as some edges extended beyond the bounds and so new points were generated to fix this. This meant that in order to reference the same Corner object, I had to look up each corner in the corners list, which had some overhead.  On top of this, work needed to be done iterating over all the edges to find the one with the same vertices to find the opposite centre.  All in all, an awful lot of iterating.

Instead of generating the regions and calculating boundary issues and all that, for now I can ignore the boundary problems and reassess later when the game design considerations are more clear.